+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change log
 
-### 0.1.0
+## 0.1.1
+#### Added
+- Check for existing tasks, launch & ruleset files in workspace.
+- Prompt to replace existing files if found.
+#### Fixes
+- Configuration settings are now implemented properly.
+
+## 0.1.0
 - Initial release